+++ cppcrypto/crypto_hash.h
@@ -22,8 +22,8 @@ public:
 	virtual void update(const uint8_t* data, size_t len) = 0;
 	virtual void final(uint8_t* hash) = 0;
 
-	virtual int hashsize() const = 0;
-	virtual int blocksize() const = 0;
+	virtual size_t hashsize() const = 0;
+	virtual size_t blocksize() const = 0;
 	virtual crypto_hash* clone() const = 0;
 
 	void hash_string(const uint8_t* data, size_t len, uint8_t* hash);
