+++ cppcrypto/groestl.h
@@ -25,8 +25,8 @@ namespace cppcrypto
 		void update(const uint8_t* data, size_t len);
 		void final(uint8_t* hash);
 
-		int hashsize() const { return 256; }
-		int blocksize() const { return 512; }
+		size_t hashsize() const { return 256; }
+		size_t blocksize() const { return 512; }
 		crypto_hash* clone() const { return new groestl256; }
 
 	private:
@@ -50,8 +50,8 @@ namespace cppcrypto
 		void update(const uint8_t* data, size_t len);
 		void final(uint8_t* hash);
 
-		int hashsize() const { return 512; }
-		int blocksize() const { return 1024; }
+		size_t hashsize() const { return 512; }
+		size_t blocksize() const { return 1024; }
 		crypto_hash* clone() const { return new groestl512; }
 
 	private:
@@ -68,14 +68,14 @@ namespace cppcrypto
 	class groestl384 : public groestl512
 	{
 	public:
-		int hashsize() const { return 384; }
+		size_t hashsize() const { return 384; }
 		crypto_hash* clone() const { return new groestl384; }
 	};
 
 	class groestl224 : public groestl256
 	{
 	public:
-		int hashsize() const { return 224; }
+		size_t hashsize() const { return 224; }
 		crypto_hash* clone() const { return new groestl224; }
 	};
