+++ cppcrypto/sha1.h
@@ -24,8 +24,8 @@ namespace cppcrypto
 		void update(const uint8_t* data, size_t len);
 		void final(uint8_t* hash);
 
-		int hashsize() const { return 160; }
-		int blocksize() const { return 512; }
+		size_t hashsize() const { return 160; }
+		size_t blocksize() const { return 512; }
 		crypto_hash* clone() const { return new sha1; }
 
 	protected:
