+++ cppcrypto/skein1024.h
@@ -23,8 +23,8 @@ namespace cppcrypto
 		void update(const uint8_t* data, size_t len);
 		void final(uint8_t* hash);
 
-		int hashsize() const { return 1024; }
-		int blocksize() const { return 1024; }
+		size_t hashsize() const { return 1024; }
+		size_t blocksize() const { return 1024; }
 		crypto_hash* clone() const { return new skein1024_1024; }
 
 	protected:
@@ -46,7 +46,7 @@ namespace cppcrypto
 	public:
 		void init();
 
-		int hashsize() const { return 512; }
+		size_t hashsize() const { return 512; }
 		crypto_hash* clone() const { return new skein1024_512; }
 	};
 
@@ -55,7 +55,7 @@ namespace cppcrypto
 	public:
 		void init();
 
-		int hashsize() const { return 384; }
+		size_t hashsize() const { return 384; }
 		crypto_hash* clone() const { return new skein1024_384; }
 	};
 
@@ -64,7 +64,7 @@ namespace cppcrypto
 	public:
 		void init();
 
-		int hashsize() const { return 256; }
+		size_t hashsize() const { return 256; }
 		crypto_hash* clone() const { return new skein1024_256; }
 	};
