+++ cppcrypto/skein256.h
@@ -23,8 +23,8 @@ namespace cppcrypto
 		void update(const uint8_t* data, size_t len);
 		void final(uint8_t* hash);
 
-		int hashsize() const { return 256; }
-		int blocksize() const { return 256; }
+		size_t hashsize() const { return 256; }
+		size_t blocksize() const { return 256; }
 		crypto_hash* clone() const { return new skein256_256; }
 
 	protected:
@@ -43,7 +43,7 @@ namespace cppcrypto
 	public:
 		void init();
 
-		int hashsize() const { return 224; }
+		size_t hashsize() const { return 224; }
 		crypto_hash* clone() const { return new skein256_224; }
 	};
 
@@ -52,7 +52,7 @@ namespace cppcrypto
 	public:
 		void init();
 
-		int hashsize() const { return 160; }
+		size_t hashsize() const { return 160; }
 		crypto_hash* clone() const { return new skein256_160; }
 	};
 
@@ -61,7 +61,7 @@ namespace cppcrypto
 	public:
 		void init();
 
-		int hashsize() const { return 128; }
+		size_t hashsize() const { return 128; }
 		crypto_hash* clone() const { return new skein256_128; }
 	};
